% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegEstBoot.R
\name{IntegEstBoot}
\alias{IntegEstBoot}
\title{Nonparametric bootstrap inference on the dose-response curve via our integral estimator.}
\usage{
IntegEstBoot(
  Y,
  X,
  t_eval = NULL,
  boot_num = 500,
  alpha = 0.95,
  h_bar = NULL,
  kernT_bar = "gaussian",
  h = NULL,
  b = NULL,
  C_h = 7,
  C_b = 3,
  print_bw = TRUE,
  degree = 2,
  deriv_ord = 1,
  kernT = "epanechnikov",
  kernS = "epanechnikov",
  parallel = TRUE,
  cores = 4
)
}
\arguments{
\item{Y}{The input n-dimensional outcome variable vector.}

\item{X}{The input n*(d+1) matrix. The first column of X stores the
treatment/exposure variables, while the other d columns are confounding variables.}

\item{t_eval}{The m-dimensional vector for evaluating the dose-response curve
(Default: t_eval = NULL. Then, t_eval = \code{X[,1]}, which consists of the observed
treatment variables.)}

\item{boot_num}{The number of bootstrapping times. (Default: boot_num = 500.)}

\item{alpha}{The confidence level of both the uniform confidence band and
pointwise confidence interval. (Default: alpha = 0.95.)}

\item{h_bar}{The bandwidth parameter for the Nadaraya-Watson conditional
CDF estimator. (Default: h_bar = NULL. Then, the Silverman's rule of thumb
is applied. See Chen et al. (2016) for details.)}

\item{kernT_bar}{The name of the kernel function for the Nadaraya-Watson
conditional CDF estimator. (Default: kernT_bar = "gaussian".)}

\item{h, b}{The bandwidth parameters for the treatment/exposure variable
and confounding variables in the local polynomial regression.
(Default: h = NULL, b = NULL. Then, the rule-of-thumb bandwidth selector
in Eq. (A1) of Yang and Tschernig (1999) is used with additional scaling
factors C_h and C_b, respectively.)}

\item{C_h, C_b}{The scaling factors for the rule-of-thumb bandwidth parameters.}

\item{print_bw}{The indicator of whether the current bandwidth parameters
should be printed to the console. (Default: print_bw = TRUE.)}

\item{degree}{Degree of local polynomials. (Default: degree = 2.)}

\item{deriv_ord}{The order of the estimated derivative of the conditional mean
outcome function. (Default: deriv_ord = 1. It shouldn't be changed in most cases.)}

\item{kernT, kernS}{The names of kernel functions for the treatment/exposure
variable and confounding variables. (Default: kernT = "epanechnikov",
kernS = "epanechnikov".)}

\item{parallel}{The indicator of whether the function should be parallel
executed. (Default: parallel = TRUE.)}

\item{cores}{The number of cores for parallel execution. (Default: cores = 6.)}
}
\value{
A list that contains four elements.
\item{m_est}{The estimated dose-response curve evaluated at points \code{t_eval}.}
\item{m_est_boot}{The estimated dose-response curve evaluated at points \code{t_eval} for all the bootstrap samples.}
\item{m_alpha}{The width of the uniform confidence band.}
\item{m_alpha_var}{The widths of the pointwise confidence bands at evaluation points \code{t_eval}.}
}
\description{
This function implements the nonparametric bootstrap inference on the
dose-response curve via our integral estimator.
}
\references{
Zhang, Y., Chen, Y.-C., and Giessing, A. (2024)
\emph{Nonparametric Inference on Dose-Response Curves Without the Positivity Condition.}
\url{https://arxiv.org/abs/2405.09003}.
}
\author{
Yikun Zhang, \email{yikunzhang@foxmail.com}
}
\keyword{bootstrap}
\keyword{curve}
\keyword{dose-response}
\keyword{estimator}
\keyword{inference}
\keyword{integral}
\keyword{on}
\keyword{our}
\keyword{the}
\keyword{through}
