% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernelFunc.R
\name{rectangular}
\alias{rectangular}
\title{The commonly used kernel functions.}
\usage{
rectangular(t)
}
\arguments{
\item{t}{The query point or numeric vector.}
}
\value{
The kernel values evaluated at the query points.
}
\description{
This file implements all the commonly used kernel functions.
}
\author{
Yikun Zhang, \email{yikunzhang@foxmail.com}
}
\keyword{common}
\keyword{functions}
\keyword{kernel}
